renamed folders using camelcase, other small changes
