+++ server/repository/assessmentCountry/readAll.ts
@@ -12,11 +12,11 @@ export const readAll = async (
 
   const assessmentName = Schemas.getName(assessment)
 
-  const xxx = await client.many<AssessmentCountry>(
+  const items = await client.many<AssessmentCountry>(
     `
         select * from ${assessmentName}.assessment_country;
     `
   )
 
-  return Objects.camelize(xxx)
+  return Objects.camelize(items)
 }
