+++ server/repository/regionGroup/readAll.ts
@@ -3,11 +3,11 @@ import { RegionGroup } from '@core/meta/regionGroup'
 import { Objects } from '@core/utils'
 
 export const readAll = async (client: BaseProtocol = DB): Promise<Array<RegionGroup>> => {
-  const xxx = await client.many<RegionGroup>(
+  const items = await client.many<RegionGroup>(
     `
         select * from public.region_group;
     `
   )
 
-  return Objects.camelize(xxx)
+  return Objects.camelize(items)
 }
