+++ server/service/index.ts
@@ -1,6 +1,6 @@
 export { AssessmentService } from './assessment'
-export { AssessmentCountryService } from './assessment_country'
-export { AssessmentRegionService } from './assessment_region'
-export { RegionGroupService } from './region_group'
+export { AssessmentCountryService } from './assessmentCountry'
+export { AssessmentRegionService } from './assessmentRegion'
+export { RegionGroupService } from './regionGroup'
 export { UserService } from './user'
 export { SettingsService } from './settings'
